pyproject.toml
[tool.poetry]
name = "gen_excel"
version = "0.1.0"
description = ""
authors = ["Николай Некрасов <nnnekrasov@edu.hse.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openpyxl = "^3.1.2"
matplotlib = "^3.7.1"
sqlalchemy = "^2.0.16"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.6"
pandas = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
